Lazy-import Textual to halve CLI startup time
autowt ls went from ~320ms to ~160ms. Textual was being imported at
module load time through three chains even for commands that don't
use TUIs. Now:

- cli.py: run_switch_tui imported inside _run_interactive_switch()
- cleanup.py: run_cleanup_tui imported inside _interactive_selection()
- config.py: ConfigApp moved to tui/config.py, imported inside
  configure_settings(). config.css moved alongside it.
